+++ AddOns/BigDebuffs/BigDebuffs.lua
@@ -337,6 +337,22 @@ else
             Curse = function() return IsUsableSpell(GetSpellInfo(374251)) end,
         },
     }
+    if WOW_PROJECT_ID == WOW_PROJECT_CATACLYSM_CLASSIC then
+        classDispel.DRUID = {
+            Magic = function() return IsPlayerSpell(88423) end,
+            Curse = true,
+            Poison = true,
+        }
+        classDispel.SHAMAN = {
+            Magic = function() return IsPlayerSpell(77130) end,
+            Curse = true,
+        }
+        classDispel.PALADIN = {
+            Magic = function() return IsPlayerSpell(53551) end,
+            Poison = true,
+            Disease = true,
+        }
+    end
     local _, class = UnitClass("player")
     BigDebuffs.dispelTypes = classDispel[class]
 end
@@ -381,6 +397,7 @@ local GetAnchor = {
                     end
                 end
             end
+            return
         end
 
         if unit and (unit:match("arena") or unit:match("arena")) then