+++ AddOns/BigDebuffs/CHANGELOG.md
@@ -1,1 +1,7 @@
-- Bug fix
+# BigDebuffs
+
+## [v15](https://github.com/jordonwow/bigdebuffs/tree/v15) (2024-05-29)
+[Full Changelog](https://github.com/jordonwow/bigdebuffs/compare/v14...v15) [Previous Releases](https://github.com/jordonwow/bigdebuffs/releases)
+
+- Fix class dispel abilities for Druid  
+    Fixes #732